refactor(server): split cloud command result store
